src/requirements.txt
streamlit==1.32.0
langchain==0.2.1
faiss-cpu==1.8.0
sentence-transformers==2.7.0
requests==2.32.2
pypdf==4.2.0
tqdm==4.66.4
# Removed geopandas and pyproj
# Simplified leafmap installation
leafmap==0.25.3